corrupted_vision/challenge/src/requirements.txt
blinker==1.8.1
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.6.0
gunicorn==22.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
limits==3.11.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
pycryptodome==3.20.0
Pygments==2.17.2
rich==13.7.1
typing_extensions==4.11.0
Werkzeug==3.0.2
wrapt==1.16.0
